+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - Expiration support (see [the documentation](https://apollographql.github.io/apollo-kotlin-normalized-cache-incubating/expiration.html) for details)
 - Compatibility with the IntelliJ plugin cache viewer (#42)
 - For consistency, `MemoryCacheFactory` and `MemoryCache` are now in the `com.apollographql.cache.normalized.memory` package 
-- Remove deprecated symbols 
+- Remove deprecated symbols
+- Add `IdCacheKeyGenerator` and `IdCacheKeyResolver` (#41)
 
 # Version 0.0.3
 _2024-09-20_